V1 Collector + API refactor  (#487)
* V1 refactor - phase 2.

* V1 refactor - phase 2.

* FeatureServiceImpl refactor.
