--- versionone-feature-collector/src/main/java/com/capitalone/dashboard/collector/RestOperationsSupplier.java
@@ -1,43 +1,0 @@
-/*************************DA-BOARD-LICENSE-START*********************************
- * Copyright 2014 CapitalOne, LLC.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *************************DA-BOARD-LICENSE-END*********************************/
-
-package com.capitalone.dashboard.collector;
-
-import com.capitalone.dashboard.util.Supplier;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestOperations;
-import org.springframework.web.client.RestTemplate;
-
-/**
- * Supplier that returns a new {@link RestTemplate}.
- */
-@Component
-public class RestOperationsSupplier implements Supplier<RestOperations> {
-	/**
-	 * Handles the REST operation HTTP connection timeout behavior
-	 * 
-	 * @return A configured REST template artifact
-	 */
-	@Override
-	public RestOperations get() {
-		HttpComponentsClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory();
-		requestFactory.setConnectTimeout(40000);
-		requestFactory.setReadTimeout(40000);
-
-		return new RestTemplate(requestFactory);
-	}
-}
