+++ versionone-feature-collector/src/main/java/com/capitalone/dashboard/model/FeatureCollector.java
@@ -1,18 +1,3 @@
-/*************************DA-BOARD-LICENSE-START*********************************
- * Copyright 2014 CapitalOne, LLC.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *************************DA-BOARD-LICENSE-END*********************************/
 
 package com.capitalone.dashboard.model;
 
@@ -21,8 +6,6 @@ import com.capitalone.dashboard.util.FeatureCollectorConstants;
 /**
  * Collector implementation for Feature that stores system configuration
  * settings required for source system data connection (e.g., API tokens, etc.)
- * 
- * @author KFK884
  */
 public class FeatureCollector extends Collector {
 	/**
