--- versionone-feature-collector/src/test/java/com/capitalone/dashboard/datafactory/versionone/test/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Houses test methods for VersionOne data connection and response building capabilities.
- * 
- * @author KFK884
- *
- */
-package com.capitalone.dashboard.datafactory.versionone.test;
